+++ packages/components/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+### [2.9.0-beta.1] - 2022-01-27
+
 #### Added
 - `:focus-visible` Content of selected Tab in `Tabs` component gets focus styling
 - improved accessibility of `Text Field Wrapper` and `Textarea Wrapper` when `maxlength` attribute is set